# docs/swagger.yaml
definitions:
...
  CreateUserInput:
    properties:
      name:
        type: string
      oauth_provider:
        type: string
      oauth_provider_data:
        additionalProperties: true
        type: object
      oauth_provider_id:
        type: string
      password:
        minLength: 8
        type: string
      profile_picture:
        type: string
      role_names:
        items:
          type: string
        minItems: 1
        type: array
      username:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - name
    - role_names
    - username
    type: object
...
  ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  ListUsersResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/User'
        type: array
    type: object
  LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  OAuthLinkRequest:
    properties:
      provider:
        type: string
      provider_user_id:
        type: string
      user_id:
        type: string
      user_info:
        additionalProperties: true
        type: object
    required:
    - provider
    - provider_user_id
    - user_id
    - user_info
    type: object
...
  Permission:
    properties:
      action:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      resource_type:
        type: string
    type: object
...
  SSOProvider:
    properties:
      issuer_url:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  SSOProvidersResponse:
    properties:
      providers:
        items:
          $ref: '#/definitions/SSOProvider'
        type: array
    type: object
...
  TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      mfa_token:
        type: string
      requires_password_change:
        type: boolean
      requires_totp:
        type: boolean
      token_type:
        type: string
    type: object
...
  UpdatePasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
    required:
    - new_password
    type: object
...
  UpdateUserInput:
    properties:
      active:
        type: boolean
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      preferences:
        additionalProperties: true
        type: object
      profile_picture:
        type: string
      role_names:
        items:
          type: string
        minItems: 1
        type: array
    type: object
  User:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      id:
        type: string
      identities:
        items:
          $ref: '#/definitions/UserIdentity'
        type: array
      must_change_password:
        type: boolean
      name:
        type: string
      preferences:
        additionalProperties: true
        type: object
      profile_picture:
        type: string
      roles:
        items:
          $ref: '#/definitions/UserRole'
        type: array
      sessions_invalidated_at:
        description: Tokens issued before this moment are rejected at validation,
          so setting it to now signs the user out of every session at once.
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  UserIdentity:
    properties:
      created_at:
        type: string
      id:
        type: string
      provider:
        type: string
      provider_data:
        additionalProperties: true
        type: object
      provider_email:
        type: string
      provider_user_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  UserRole:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/Permission'
        type: array
    type: object
...
  ValidationErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  ValidationErrorResponse:
    properties:
      error:
        type: string
      fields:
        items:
          $ref: '#/definitions/ValidationErrorDetail'
        type: array
    type: object
...
info:
  contact: {}
  description: API for interacting with Marmot
  license:
    name: MIT
    url: https://opensource.org/license/MIT
  title: Marmot API
  version: "0.1"
paths:
...
  /api/v1/sso-providers:
    get:
      description: Read-only view of SSO providers wired via server config. Editing
        is done in config.yaml.
      operationId: getSsoProviders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SSOProvidersResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List configured SSO providers (admin)
      tags:
      - auth
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Get a list of users with optional filtering
      operationId: getUsers
      parameters:
      - default: 50
        description: Number of items to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      - description: Search query for username or email
        in: query
        name: query
        type: string
      - collectionFormat: csv
        description: Filter by role IDs
        in: query
        items:
          type: string
        name: role_ids
        type: array
      - description: Filter by active status
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ListUsersResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user in the system
      operationId: postUsers
      parameters:
      - description: User creation request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/CreateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ValidationErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Create a new user
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user from the system
      operationId: deleteUsersID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific user
      operationId: getUsersID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get a user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user information
      operationId: putUsersID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/UpdateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ValidationErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update a user
      tags:
      - users
...
  /api/v1/users/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user with username/email and password
      operationId: postUsersLogin
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Login user
      tags:
      - users
...
  /api/v1/users/oauth/link:
    post:
      consumes:
      - application/json
      description: Link an OAuth account to an existing user
      operationId: postUsersOauthLink
      parameters:
      - description: OAuth account link request
        in: body
        name: link
        required: true
        schema:
          $ref: '#/definitions/OAuthLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Link OAuth account
      tags:
      - users
  /api/v1/users/oauth/unlink/{id}/{provider}:
    delete:
      consumes:
      - application/json
      description: Unlink an OAuth account from a user
      operationId: deleteUsersOauthUnlinkIDProvider
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: OAuth provider
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Unlink OAuth account
      tags:
      - users
...
  /api/v1/users/sign-out-all/{id}:
    post:
      description: Rejects every token issued to the user before now. Their password
        and API keys are untouched, so this is the response to a leaked or overexposed
        token rather than a lost credential.
      operationId: postUsersIDSignOutAll
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Sign a user out of all sessions
      tags:
      - users
...
  /api/v1/users/update-password:
    post:
      consumes:
      - application/json
      description: Update current user's password
      operationId: postUsersUpdatePassword
      parameters:
      - description: Password update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update user password
      tags:
      - users
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"