# internal/api/auth/config.go
package auth

import (
	"net/http"

	"github.com/marmotdata/marmot/internal/api/v1/common"
	coreauth "github.com/marmotdata/marmot/internal/core/auth"
)

type AuthConfig struct {
	TOTPEnabled      bool     `json:"totp_enabled,omitempty"`
	EnabledProviders []string `json:"enabled_providers"`
} // @name AuthConfig

// @Summary Get auth configuration
// @Description Returns the enabled auth providers without sensitive data
// @Tags auth
// @Produce json
// @Success 200 {object} AuthConfig
// @ID getAuthProviders
// @Router /auth-providers [get]
func (h *Handler) getAuthConfig(w http.ResponseWriter, r *http.Request) {
	config := AuthConfig{
		EnabledProviders: h.oauthManager.GetProviderNames(),
		TOTPEnabled:      h.config.Auth.TOTP.Enabled,
	}

	common.RespondJSON(w, http.StatusOK, config)
}

// SSOProvider is an admin-facing view of a configured SSO provider.
type SSOProvider struct {
	Type      string `json:"type"`
	Name      string `json:"name"`
	IssuerURL string `json:"issuer_url,omitempty"`
} // @name SSOProvider

type ssoProvidersResponse struct {
	Providers []SSOProvider `json:"providers"`
} // @name SSOProvidersResponse

// @Summary List configured SSO providers (admin)
// @Description Read-only view of SSO providers wired via server config. Editing is done in config.yaml.
// @Tags auth
// @Produce json
// @Security ApiKeyAuth
// @Security BearerAuth
// @Success 200 {object} ssoProvidersResponse
// @ID getSsoProviders
// @Router /api/v1/sso-providers [get]
func (h *Handler) getSSOProviders(w http.ResponseWriter, r *http.Request) {
	out := ssoProvidersResponse{Providers: []SSOProvider{}}
	for _, p := range h.oauthManager.GetProviders() {
		item := SSOProvider{Type: p.Type(), Name: p.Name()}
		if ip, ok := p.(coreauth.IssuerProvider); ok {
			item.IssuerURL = ip.IssuerURL()
		}
		out.Providers = append(out.Providers, item)
	}
	common.RespondJSON(w, http.StatusOK, out)
}