# internal/api/v1/users/auth.go
package users

import (
	"encoding/json"
	"errors"
	"net/http"
	"strings"

	"github.com/marmotdata/marmot/internal/api/v1/common"
	"github.com/marmotdata/marmot/internal/core/mfa"
	"github.com/marmotdata/marmot/internal/core/user"
	"github.com/rs/zerolog/log"
)

type UpdatePasswordRequest struct {
	NewPassword string `json:"new_password" validate:"required,min=8"`
} // @name UpdatePasswordRequest

type LoginRequest struct {
	Username string `json:"username" validate:"required"`
	Password string `json:"password" validate:"required"`
} // @name LoginRequest

type TokenResponse struct {
	RequiresTOTP           bool   `json:"requires_totp,omitempty"`
	MFAToken               string `json:"mfa_token,omitempty"`
	AccessToken            string `json:"access_token,omitempty"`
	TokenType              string `json:"token_type,omitempty"`
	ExpiresIn              int64  `json:"expires_in,omitempty"`
	RequiresPasswordChange bool   `json:"requires_password_change"`
} // @name TokenResponse

type OAuthLinkRequest struct {
	UserID         string                 `json:"user_id" validate:"required"`
	Provider       string                 `json:"provider" validate:"required"`
	ProviderUserID string                 `json:"provider_user_id" validate:"required"`
	UserInfo       map[string]interface{} `json:"user_info" validate:"required"`
} // @name OAuthLinkRequest

// @Summary Login user
// @Description Authenticate a user with username/email and password
// @Tags users
// @Accept json
// @Produce json
// @Param credentials body LoginRequest true "Login credentials"
// @Success 200 {object} TokenResponse
// @Failure 401 {object} common.ErrorResponse
// @Failure 400 {object} common.ErrorResponse
// @ID postUsersLogin
// @Router /api/v1/users/login [post]
func (h *Handler) login(w http.ResponseWriter, r *http.Request) {
	var input LoginRequest
	if err := json.NewDecoder(r.Body).Decode(&input); err != nil {
		common.RespondError(w, http.StatusBadRequest, "Invalid request body")
		return
	}

	authenticatedUser, err := h.userService.Authenticate(r.Context(), input.Username, input.Password)
	if err != nil {
		// errors.Is, not ==: these are bare today, but the rest of core/user
		// wraps its sentinels and a value comparison would silently rot.
		if errors.Is(err, user.ErrInvalidPassword) || errors.Is(err, user.ErrUnauthorized) {
			common.RespondError(w, http.StatusUnauthorized, "Invalid credentials")
			return
		}
		log.Error().Err(err).Str("username", input.Username).Msg("Authentication failed")
		common.RespondError(w, http.StatusInternalServerError, "Authentication failed")
		return
	}

	h.completeLocalLogin(w, r, authenticatedUser)
}

func (h *Handler) completeLocalLogin(w http.ResponseWriter, r *http.Request, u *user.User) {
	if h.config.Auth.TOTP.Enabled {
		if h.mfa == nil {
			h.respondMFAError(w, mfa.ErrUnavailable)
			return
		}
		status, err := h.mfa.Status(r.Context(), u.ID)
		if err != nil {
			h.respondMFAError(w, err)
			return
		}
		purpose := ""
		if u.MustChangePassword {
			purpose = mfa.PurposePasswordChange
		} else if status.Enabled {
			purpose = mfa.PurposeTOTP
		}
		if purpose != "" {
			token, err := h.mfa.IssueChallenge(r.Context(), u.ID, purpose)
			if err != nil {
				h.respondMFAError(w, err)
				return
			}
			w.Header().Set("Cache-Control", "no-store")
			common.RespondJSON(w, http.StatusOK, TokenResponse{MFAToken: token, RequiresTOTP: purpose == mfa.PurposeTOTP, RequiresPasswordChange: u.MustChangePassword})
			return
		}
	}
	h.issueSession(w, r, u)
}

func (h *Handler) issueSession(w http.ResponseWriter, r *http.Request, authenticatedUser *user.User) {
	theme := ""
	if authenticatedUser.Preferences != nil {
		if themeVal, ok := authenticatedUser.Preferences["theme"].(string); ok {
			theme = themeVal
		}
	}

	extraClaims := map[string]interface{}{
		"theme": theme,
	}

	token, err := h.authService.GenerateToken(r.Context(), authenticatedUser, extraClaims)
	if err != nil {
		log.Error().Err(err).Str("user_id", authenticatedUser.ID).Msg("Failed to generate token")
		common.RespondError(w, http.StatusInternalServerError, "Failed to generate token")
		return
	}

	response := TokenResponse{
		AccessToken:            token,
		TokenType:              "Bearer",
		ExpiresIn:              24 * 60 * 60,
		RequiresPasswordChange: authenticatedUser.MustChangePassword,
	}

	common.RespondJSON(w, http.StatusOK, response)
}

// @Summary Link OAuth account
// @Description Link an OAuth account to an existing user
// @Tags users
// @Accept json
// @Produce json
// @Param link body OAuthLinkRequest true "OAuth account link request"
// @Security ApiKeyAuth
// @Security BearerAuth
// @Success 200 "OK"
// @Failure 400 {object} common.ErrorResponse
// @ID postUsersOauthLink
// @Router /api/v1/users/oauth/link [post]
func (h *Handler) linkOAuthAccount(w http.ResponseWriter, r *http.Request) {
	var input OAuthLinkRequest
	if err := json.NewDecoder(r.Body).Decode(&input); err != nil {
		common.RespondError(w, http.StatusBadRequest, "Invalid request body")
		return
	}

	err := h.userService.LinkOAuthAccount(r.Context(), input.UserID, input.Provider, input.ProviderUserID, input.UserInfo)
	if err != nil {
		log.Error().Err(err).
			Str("user_id", input.UserID).
			Str("provider", input.Provider).
			Msg("Failed to link OAuth account")
		common.RespondError(w, http.StatusInternalServerError, "Failed to link OAuth account")
		return
	}

	w.WriteHeader(http.StatusOK)
}

// @Summary Unlink OAuth account
// @Description Unlink an OAuth account from a user
// @Tags users
// @Accept json
// @Produce json
// @Param id path string true "User ID"
// @Param provider path string true "OAuth provider"
// @Security ApiKeyAuth
// @Security BearerAuth
// @Success 204 "No Content"
// @Failure 400 {object} common.ErrorResponse
// @ID deleteUsersOauthUnlinkIDProvider
// @Router /api/v1/users/oauth/unlink/{id}/{provider} [delete]
func (h *Handler) unlinkOAuthAccount(w http.ResponseWriter, r *http.Request) {
	parts := strings.Split(strings.TrimPrefix(r.URL.Path, "/api/v1/users/oauth/unlink/"), "/")
	if len(parts) != 2 {
		common.RespondError(w, http.StatusBadRequest, "Invalid path")
		return
	}

	userID := parts[0]
	provider := parts[1]

	err := h.userService.UnlinkOAuthAccount(r.Context(), userID, provider)
	if err != nil {
		log.Error().Err(err).
			Str("user_id", userID).
			Str("provider", provider).
			Msg("Failed to unlink OAuth account")
		common.RespondError(w, http.StatusInternalServerError, "Failed to unlink OAuth account")
		return
	}

	w.WriteHeader(http.StatusNoContent)
}

// @Summary Sign a user out of all sessions
// @Description Rejects every token issued to the user before now. Their password and API keys are untouched, so this is the response to a leaked or overexposed token rather than a lost credential.
// @Tags users
// @Produce json
// @Param id path string true "User ID"
// @Security ApiKeyAuth
// @Security BearerAuth
// @Success 204
// @Failure 404 {object} common.ErrorResponse
// @ID postUsersIDSignOutAll
// @Router /api/v1/users/sign-out-all/{id} [post]
func (h *Handler) signOutAllSessions(w http.ResponseWriter, r *http.Request) {
	id := r.PathValue("id")
	if id == "" {
		http.NotFound(w, r)
		return
	}

	if err := h.userService.InvalidateSessions(r.Context(), id); err != nil {
		if errors.Is(err, user.ErrUserNotFound) {
			common.RespondError(w, http.StatusNotFound, "User not found")
			return
		}
		log.Error().Err(err).Str("user_id", id).Msg("Failed to invalidate sessions")
		common.RespondError(w, http.StatusInternalServerError, "Failed to sign out sessions")
		return
	}

	w.WriteHeader(http.StatusNoContent)
}

// @Summary Update user password
// @Description Update current user's password
// @Tags users
// @Accept json
// @Produce json
// @Param request body UpdatePasswordRequest true "Password update request"
// @Security ApiKeyAuth
// @Security BearerAuth
// @Success 200 {object} TokenResponse
// @Failure 400 {object} common.ErrorResponse
// @Failure 401 {object} common.ErrorResponse
// @ID postUsersUpdatePassword
// @Router /api/v1/users/update-password [post]
func (h *Handler) updatePassword(w http.ResponseWriter, r *http.Request) {
	var input UpdatePasswordRequest
	if err := json.NewDecoder(r.Body).Decode(&input); err != nil {
		common.RespondError(w, http.StatusBadRequest, "Invalid request body")
		return
	}

	usr, ok := common.GetAuthenticatedUser(r.Context())
	if !ok {
		common.RespondError(w, http.StatusUnauthorized, "Authentication required")
		return
	}

	updatedUser, err := h.userService.UpdatePassword(r.Context(), usr.ID, input.NewPassword)
	if err != nil {
		if respondUserError(w, err) {
			return
		}
		log.Error().Err(err).Str("user_id", usr.ID).Msg("Failed to update password")
		common.RespondError(w, http.StatusInternalServerError, "Failed to update password")
		return
	}

	theme := ""
	if updatedUser.Preferences != nil {
		if themeVal, ok := updatedUser.Preferences["theme"].(string); ok {
			theme = themeVal
		}
	}

	extraClaims := map[string]interface{}{
		"theme": theme,
	}

	token, err := h.authService.GenerateToken(r.Context(), updatedUser, extraClaims)
	if err != nil {
		common.RespondError(w, http.StatusInternalServerError, "Failed to generate token")
		return
	}

	response := TokenResponse{
		AccessToken: token,
		TokenType:   "Bearer",
		ExpiresIn:   24 * 60 * 60,
	}

	common.RespondJSON(w, http.StatusOK, response)
}